Rename all files according to flutter coding style and remove warnigngs from dart analysis
