+++ lib/ui/notifications/ct_loader.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 class CTLoader extends StatelessWidget {
-
   final double width;
   final double height;
 
   CTLoader(this.width, this.height);
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -21,6 +21,4 @@ class CTLoader extends StatelessWidget {
       ),
     );
   }
-
-
 }
