+++ lib/utils/app_localization.dart
@@ -5,30 +5,32 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 class AppLocalization {
-
   static final AppLocalization _singleton = new AppLocalization._internal();
+
   AppLocalization._internal();
+
   static AppLocalization get instance => _singleton;
 
   Map<dynamic, dynamic> _localisedValues;
   Map<dynamic, dynamic> _englishValues;
 
   Future<AppLocalization> load(Locale locale) async {
-    
-    String jsonContent =
-      await rootBundle.loadString("assets/locales/${locale != null ? locale.languageCode : "en"}.json");
+    String jsonContent = await rootBundle.loadString(
+        "assets/locales/${locale != null ? locale.languageCode : "en"}.json");
     _localisedValues = json.decode(jsonContent);
-    if(locale != null && locale.languageCode != 'en') {
+    if (locale != null && locale.languageCode != 'en') {
       String jsonEnglishContent =
-      await rootBundle.loadString("assets/locales/en.json");
-    _englishValues = json.decode(jsonEnglishContent);
+          await rootBundle.loadString("assets/locales/en.json");
+      _englishValues = json.decode(jsonEnglishContent);
     }
-    
+
     return this;
   }
 
   static String text(String key) {
-    return instance._localisedValues[key] ?? instance._englishValues[key] ?? key;
+    return instance._localisedValues[key] ??
+        instance._englishValues[key] ??
+        key;
   }
 }
 
@@ -39,7 +41,7 @@ class AppLocalizationsDelegate extends LocalizationsDelegate<AppLocalization> {
   bool isSupported(Locale locale) => ['en', 'es'].contains(locale.languageCode);
 
   @override
-  Future<AppLocalization> load(Locale locale)  {
+  Future<AppLocalization> load(Locale locale) {
     return AppLocalization.instance.load(locale);
   }
